requirements.txt
Jinja2
numpy
openai==1.97.0
pandas
Requests
tqdm
requests
retrying
sympy
math-verify
json_repair
aiohttp
chardet
antlr4-python3-runtime==4.11
aiofiles==24.1.0
google_search_results==2.4.2
matplotlib==3.8.4
serpapi==0.1.5
together==1.4.6
jsonlines
human-eval
qiskit
qiskit_ibm_runtime
matplotlib
qiskit_aer
networkx
qiskit_ibm_transpiler
anthropic==0.57.1
openpyxl==3.1.5
pydantic==2.11.7
volcengine_python_sdk==1.0.115
wolframclient==1.4.0
rdkit==2025.3.2
selfies==2.2.0
torch==2.7.1
tenacity    
